Refactor ranking-chart; add ranking-chart-plus
This commit extends existing ranking-chart component to add a ranking-chart-plus component, which simple overlays a secondary statistic on top of the existing visualization. This in turn required some refactoring of the ranking-chart component. It introduces no breaking changes i nthe current API. It closes #306 by using the correct classes in the d3 selectAll. It addresses #301 by adding in the new comparison chart component. That component can be embedded like this:

```
{{ranking-chart-plus
  column='cd_tot_bldgs'
  overlayColumn='current_fp_bldg'
  numeral_format='0,0'
  borocd=model.borocd}}
```

The key-indicators component is changed, but only trivially - it removes an unnecessary then method.