593: land-use-map -- update fill-color & tooltip logic to allow for undefined values to be in Other category
Some rows in mappluto have a null value for the `landuse` field. When
mappluto is joined with support_landuse_lookup table, this means that
these rows will be styled (color and tooltip) as undefined on the
land-use-map. This PR updates the fill-color and tooltip description
logic to account for these undefined values by considering them as
being in the 'Other' category.